backend/requirements.txt
anyio==3.6.2
asyncinotify==4.0.2
click==8.1.3
colorcet==3.0.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
h11==0.14.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
param==1.13.0
Pillow==9.5.0
pyct==0.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tzdata==2023.3
uvicorn==0.22.0
